feat: 논문 업로드 페이지 component 분리
